Delete k8s_instance_manager.py
